+++ src/OSharp.Hosting.Apis/Areas/Admin/Controllers/Systems/MenuController.cs
@@ -72,7 +72,7 @@ public class MenuController : AdminApiControllerBase
     [DependOnFunction(nameof(Read))]
     [UnitOfWork]
     [Description("删除")]
-    public async Task<AjaxResult> Delete(int[] ids)
+    public async Task<AjaxResult> Delete(long[] ids)
     {
         Check.NotNull(ids, nameof(ids));
