+++ src/OSharp.Hosting.Core/Authorization/Entities/EntityRole.cs
@@ -16,7 +16,7 @@ namespace OSharp.Hosting.Authorization.Entities;
 /// 实体：数据角色信息
 /// </summary>
 [Description("数据角色信息")]
-public class EntityRole : EntityRoleBase<int>
+public class EntityRole : EntityRoleBase<long>
 {
     /// <summary>
     /// 获取或设置 所属角色信息
