+++ src/OSharp.Hosting.Core/Authorization/FunctionAuthCache.cs
@@ -16,7 +16,7 @@ namespace OSharp.Hosting.Authorization;
 /// <summary>
 /// 功能权限缓存
 /// </summary>
-public class FunctionAuthCache : FunctionAuthCacheBase<ModuleFunction, ModuleRole, ModuleUser, Function, Module, int, Role, int, User, int>
+public class FunctionAuthCache : FunctionAuthCacheBase<ModuleFunction, ModuleRole, ModuleUser, Function, Module, long, Role, long, User, long>
 {
     /// <summary>
     /// 初始化一个<see cref="FunctionAuthCacheBase{TModuleFunction, TModuleRole, TModuleUser, TFunction, TModule, TModuleKey,TRole, TRoleKey, TUser, TUserKey}"/>类型的新实例
