+++ src/OSharp.Hosting.Core/Authorization/ModuleHandler.cs
@@ -16,7 +16,7 @@ namespace OSharp.Hosting.Authorization;
 /// <summary>
 /// 模块信息处理器
 /// </summary>
-public class ModuleHandler : ModuleHandlerBase<Module, ModuleInputDto, int, ModuleFunction>
+public class ModuleHandler : ModuleHandlerBase<Module, ModuleInputDto, long, ModuleFunction>
 {
     /// <summary>
     /// 初始化一个<see cref="ModuleHandlerBase{TModule, TModuleInputDto, TModuleKey, TModuleFunction}"/>类型的新实例
