+++ src/OSharp.Hosting.Core/Identity/Entities/Organization.cs
@@ -14,7 +14,7 @@ namespace OSharp.Hosting.Identity.Entities;
 /// </summary>
 [Description("组织机构信息")]
 [TableNamePrefix("Identity")]
-public class Organization : EntityBase<int>
+public class Organization : EntityBase<long>
 {
     /// <summary>
     /// 获取或设置 名称
@@ -32,5 +32,5 @@ public class Organization : EntityBase<int>
     /// 获取或设置 父组织机构
     /// </summary>
     [DisplayName("父组织机构编号")]
-    public int? ParentId { get; set; }
+    public long? ParentId { get; set; }
 }
