+++ src/OSharp.Hosting.Core/Identity/Entities/RoleClaim.cs
@@ -14,7 +14,7 @@ namespace OSharp.Hosting.Identity.Entities;
 /// 实体类：角色声明信息
 /// </summary>
 [Description("角色声明信息")]
-public class RoleClaim : RoleClaimBase<int, int>
+public class RoleClaim : RoleClaimBase<long, long>
 {
     /// <summary>
     /// 获取或设置 所属角色信息
