+++ src/OSharp.Hosting.Core/Identity/IdentityPack.cs
@@ -20,7 +20,7 @@ namespace OSharp.Hosting.Identity;
 /// 身份认证模块，此模块必须在MVC模块之前启动
 /// </summary>
 [Description("身份认证模块")]
-public class IdentityPack : IdentityPackBase<UserStore, RoleStore, User, int, UserClaim, int, Role, int>
+public class IdentityPack : IdentityPackBase<UserStore, RoleStore, User, long, UserClaim, long, Role, long>
 {
     /// <summary>
     /// 将模块服务添加到依赖注入服务容器中
