+++ src/OSharp.Hosting.EntityConfiguration/Identity/RoleClaimConfiguration.cs
@@ -12,7 +12,7 @@ using OSharp.Hosting.Identity.Entities;
 
 namespace OSharp.Hosting.EntityConfiguration.Identity;
 
-public partial class RoleClaimConfiguration : EntityTypeConfigurationBase<RoleClaim, int>
+public partial class RoleClaimConfiguration : EntityTypeConfigurationBase<RoleClaim, long>
 {
     /// <summary>
     /// 重写以实现实体类型各个属性的数据库配置
